Ignore TypeScript error for react-scroll-trigger
Description:
This commit includes a workaround for a persistent TypeScript
error with the react-scroll-trigger package. Despite attempts
to resolve the issue, the error could not be fixed. As a result,
the TypeScript error has been ignored to allow for successful builds
on Vercel. This ensures the continued functionality of the website
while a more permanent solution is sought.